refactor: loader, actionを分離
