+++ app/.server/actions/index.ts
@@ -1,0 +1,3 @@
+export * from "./friendsAction";
+export * from "./friendsRequestsAction";
+export * from "./settingsAction";
