+++ app/.server/loaders/index.ts
@@ -1,0 +1,3 @@
+export * from "./authLoader";
+export * from "./friendsRequestsLoader";
+export * from "./settingsLoader";
